+++ inc/config.php
@@ -59,7 +59,7 @@ $ip_suffix = ( defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ) ? '' : '.min';
 //   'login' => [],
 //
 //   // Inline scripts: [ 'handle' => [ 'src' => text/function, 'position' => 'before|after' ] ]
-//   'inline'	=> [],
+//   'inline' => [],
 //
 //   // Localize scripts: [ 'handle' => [ 'name' => name, trans => function/path_url ] ]
 //   'local' => [
@@ -74,7 +74,9 @@ $ip_suffix = ( defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ) ? '' : '.min';
 //   ],
 //
 //   // Script attributes ( defer, async, integrity ) : [ 'route' => [ [ 'handle', key ], ...] ]
-//   'attr' => []
+//   'attr' => [
+//	   'integrity' => []
+//   ]
 // ];
 //----------------------------------------------------------------------
 
@@ -151,7 +153,7 @@ add_filter( 'ipress_scripts', function() use( $ip_suffix ) {
 //   'print' => [],
 //
 //   // Theme styles: [ 'handle' => [ 'path_url', (array)dependencies, 'version', 'media' ] ... ]
-//   'custom'  => [
+//   'custom' => [
 //     'ipress' => [ IPRESS_URL . '/style.css', [], NULL ]
 //   ],
 //
@@ -159,7 +161,9 @@ add_filter( 'ipress_scripts', function() use( $ip_suffix ) {
 //   'inline' => []
 //
 //   // Style attributes ( defer, async, integrity ): [ 'route' => [ [ 'handle', key ], ...] ]
-//   'attr' 	=> []
+//   'attr' => [
+//	   'integrity' => []
+//   ]
 // ];
 //----------------------------------------------------------------------
 
@@ -303,6 +307,11 @@ add_filter( 'ipress_media_images', function () {
 	];
 } );
 
+// Set post-thumbnail support for cpt, override at lower priority
+add_filter( 'ipress_post_thumbnails_post_types', function() {
+	return [ 'post' ];
+} );
+
 //----------------------------------------------
 //	Sidebars Configuration
 //----------------------------------------------
@@ -323,6 +332,11 @@ add_filter( 'ipress_default_sidebars', function() {
 	];
 } );
 
+// Set sidebar defaults, for wrapping widgets, override at lower priority
+add_filter( 'ipress_sidebar_defaults', function( $defaults ) {
+	return $defaults; // [ 'before_widget', 'after_widget', 'before_title', 'after_title', 'class' ]
+} );
+
 // Generate footer sidebars, override at lower priority
 add_filter( 'ipress_footer_sidebars', function() {
 	return []; // or, false for numerical sidebars
@@ -340,6 +354,9 @@ add_filter ( 'ipress_widgets', function() {
 //----------------------------------------------
 //	Custom Hooks & Filters
 //----------------------------------------------
+
+// Enable or disable front page hero section, use '__return_false'to disable
+add_filter( 'ipress_custom_hero', '__return_true' );
 
 //----------------------------------------------
 //	Shortcode Configuration